feat(editor): guardado automático de edificios al salir del modo edición
- Se implementó guardado automático al presionar F10 o Esc para salir del modo editor
- Se reorganizó la lógica para asegurar que se guarden las posiciones correctas antes de limpiar el estado
- Se actualizó editor_events.py para integrar correctamente el flujo de guardado
- Se mejoró la función save_buildings_to_json para redondear coordenadas y manejar errores
- Se verificó y completó la implementación de load_buildings_from_json para cargar desde el archivo generado
- Se agregó validación visual por consola del guardado
